+++ R/bigrquery.R
@@ -2,7 +2,7 @@
 #' \describe{
 #'   \item{`bigrquery.quiet`}{Verbose output during processing? The default
 #'   value, `NA`, turns on verbose output for queries that run longer than
-#'   two seconds.  Use `TRUE` for immediate verbose output, `FALSE`
+#'   two seconds.  Use `FALSE` for immediate verbose output, `TRUE`
 #'   for quiet operation.}
 #'
 #'   \item{`bigrquery.page.size`}{Default page size for fetching data,
@@ -13,5 +13,6 @@
 #' @importFrom jsonlite unbox
 #' @importFrom Rcpp sourceCpp
 #' @importFrom tibble tibble
+#' @keywords internal
 #' @useDynLib bigrquery, .registration = TRUE
 "_PACKAGE"