+++ R/bq-perform.R
@@ -70,7 +70,7 @@ bq_perform_extract <- function(x,
     configuration = list(
       extract = list(
         sourceTable = tableReference(x),
-        destinationUris = destination_uris,
+        destinationUris = as.list(destination_uris),
         destinationFormat = unbox(destination_format),
         compression = unbox(compression),
         printHeader = unbox(print_header)
@@ -97,7 +97,7 @@ bq_perform_extract <- function(x,
 #'     table.
 #'   * "CREATE_NEVER": The table must already exist. If it does not, a
 #'     'notFound' error is returned in the job result.
-#' @param write_dispoition Specifies the action that occurs if the
+#' @param write_disposition Specifies the action that occurs if the
 #'   destination table already exists. The following values are supported:
 #'
 #'   * "WRITE_TRUNCATE": If the table already exists, BigQuery overwrites the
@@ -195,7 +195,7 @@ bq_perform_load <- function(x,
   assert_that(is.string(billing))
 
   load <- list(
-    sourceUris = source_uris,
+    sourceUris = as.list(source_uris),
     sourceFormat = unbox(source_format),
     destinationTable = tableReference(x),
     createDisposition = unbox(create_disposition),