+++ R/bq-request.R
@@ -116,7 +116,8 @@ bq_delete <- function(url, ..., query = NULL, token = get_access_cred()) {
 
 #' @importFrom httr POST add_headers config
 bq_post <- function(url, body, ..., query = NULL, token = get_access_cred()) {
-  json <- jsonlite::toJSON(body, pretty = TRUE)
+  json <- jsonlite::toJSON(body, pretty = TRUE, auto_unbox = TRUE)
+
   req <- POST(
     paste0(base_url, url),
     body = json,
@@ -131,7 +132,7 @@ bq_post <- function(url, body, ..., query = NULL, token = get_access_cred()) {
 
 #' @importFrom httr PATCH add_headers config
 bq_patch <- function(url, body, ..., query = NULL, token = get_access_cred()) {
-  json <- jsonlite::toJSON(body, pretty = TRUE)
+  json <- jsonlite::toJSON(body, pretty = TRUE, auto_unbox = TRUE)
   req <- PATCH(
     paste0(base_url, url),
     body = json,