+++ tests/testthat/test-gs-object.R
@@ -1,6 +1,7 @@
 context("test-gs-object.R")
 
 test_that("can delete objects", {
+  skip_if_no_auth()
   tb <- bq_table(bq_test_project(), "basedata", "mtcars")
 
   gs <- gs_test_object()