+++ NEWS.md
@@ -1,5 +1,8 @@
 # bigrquery 1.0.0.9000
 
+* `bq_table_download()` correctly computes page ranges if both `max_results`
+  and `start_index` are supplied (#248)
+
 * `as.character()` now translated to `SAFE_CAST(x AS STRING)` (#268).
 
 # bigrquery 1.0.0