+++ NEWS.md
@@ -1,6 +1,6 @@
 # bigrquery 1.0.0.9000
 
-* In parse_bq_type(), x now evaluates 'NUMERIC' as 'BQ_FLOAT' rather than 
+* In parse_bq_type(), x now evaluates 'NUMERIC' as 'BQ_FLOAT' rather than
   throwing error. (@paulsendavidjay, #282)
 
 * `bq_job()` tracks location so bigrquery now works painlessly with non-US/EU